move to fluentValidation

+++ Src/ScheduleService.Application.Handler/Handlers/Cities/CityDeleteHandler.cs
@@ -1,5 +1,6 @@
 ﻿using ScheduleService.Application.Shared;
 using ScheduleService.Application.Shared.Resources;
+using ScheduleService.Application.Validator.Validators.Cities;
 using ScheduleService.Domain.Command.Commands.Cities;
 using ScheduleService.Domain.Handler.Handlers;
 using ScheduleService.Domain.Handler.Handlers.Cities;
@@ -7,27 +8,28 @@ using ScheduleService.Domain.Handler.Repositories.Cities;
 
 namespace ScheduleService.Application.Handler.Handlers.Cities;
 
-internal class CityDeleteHandler : HandlerBase<CityDeleteCommand, CustomResultData>, ICityDeleteHandler
+internal class CityDeleteHandler : RequestHandler<CityDeleteCommand, CustomResultData>, ICityDeleteHandler
 {
     private readonly ICityDeleteRepository _repository;
 
-    public CityDeleteHandler(
-        IHandlerBus handlerBus,
-        ICityDeleteRepository repository) : base(handlerBus)
+    public CityDeleteHandler(ICityDeleteRepository repository)
     {
         _repository = repository;
     }
 
-    public async override Task<CustomResultData> HandleExecution(CityDeleteCommand request, CancellationToken cancellationToken)
+    public async override Task<CustomResultData> Handle(CityDeleteCommand request, CancellationToken cancellationToken)
     {
+        if (!Validate<CityDeleteValidator>(request))
+            return InvalidResponse();
+
         if (!await _repository.CheckIfExistByIdAsync(request.Id))
             AddError(ValidationResource.EntityNotFound);
 
         if (IsInvalid)
-            return InvalidResponseAsync();
+            return InvalidResponse();
 
         await _repository.DeleteAsync(request.Id);
 
-        return ValidResponseAsync();
+        return ValidResponse();
     }
 }
